Separate fragments for search input and results
