+++ src/main/java/pl/edu/zut/mad/schedule/search/SearchActivity.kt
@@ -2,50 +2,19 @@ package pl.edu.zut.mad.schedule.search
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_search.*
 import pl.edu.zut.mad.schedule.R
-import pl.edu.zut.mad.schedule.data.model.ui.Lesson
-import pl.edu.zut.mad.schedule.util.app
-import javax.inject.Inject
 
-internal class SearchActivity : AppCompatActivity(), SearchMvp.View {
-
-    @Inject internal lateinit var presenter: SearchMvp.Presenter
-    private val lessonsAdapter = LessonsAdapter()
+internal class SearchActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_search)
-        init()
+        startSearchFragment()
     }
 
-    override fun getTeacherName() = teacherNameInputView.text.toString()
-
-    override fun getTeacherSurname() = teacherSurnameInputView.text.toString()
-
-    override fun getFacultyAbbreviation() = facultyAbbreviationInputView.text.toString()
-
-    override fun getSubject() = subjectInputView.text.toString()
-
-    override fun onScheduleDownloaded(lessons: List<Lesson>) = lessonsAdapter.setLessons(lessons)
-
-    private fun init() {
-        initInjections()
-        initViews()
-    }
-
-    private fun initInjections() = app.component
-        .plus(SearchModule(this))
-        .inject(this)
-
-    private fun initViews() {
-        lessonsListView.adapter = lessonsAdapter
-        searchButton.setOnClickListener {
-            presenter.onSearch()
-        }
-        teacherNameInputView.setText("Piotr")
-        teacherSurnameInputView.setText("Piela")
-        facultyAbbreviationInputView.setText("WI")
-        subjectInputView.setText("Modelowanie i symulacja systemów")
+    private fun startSearchFragment() {
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.searchMainContainer, SearchFragment())
+            .commit()
     }
 }
