--- src/main/java/pl/edu/zut/mad/schedule/search/SearchByTeacherFragment.kt
@@ -1,16 +1,0 @@
-package pl.edu.zut.mad.schedule.search
-
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-import pl.edu.zut.mad.schedule.R
-
-class SearchByTeacherFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_search_by_teacher, container, false)
-    }
-}
