+++ src/main/java/pl/edu/zut/mad/schedule/search/SearchResultsFragment.kt
@@ -1,0 +1,22 @@
+package pl.edu.zut.mad.schedule.search
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import pl.edu.zut.mad.schedule.R
+
+class SearchResultsFragment : Fragment() {
+
+
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View =
+        inflater.inflate(R.layout.fragment_search_results, container, false)
+
+    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val lessonsAdapter = LessonsAdapter()
+
+    }
+}
