+++ src/main/res/layout/activity_search.xml
@@ -1,93 +1,7 @@
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/searchMainContainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="10dp"
-    tools:context="pl.edu.zut.mad.schedule.search.SearchActivity">
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/teacherNameLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toTopOf="@id/teacherSurnameLayout"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <android.support.design.widget.TextInputEditText
-            android:id="@+id/teacherNameInputView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/teacher_name"
-            android:inputType="text" />
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/teacherSurnameLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/teacherNameLayout">
-
-        <android.support.design.widget.TextInputEditText
-            android:id="@+id/teacherSurnameInputView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/teacher_surname"
-            android:inputType="text" />
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/facultyLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:inputType="text"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/teacherSurnameLayout">
-
-        <android.support.design.widget.TextInputEditText
-            android:id="@+id/facultyAbbreviationInputView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/faculty" />
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/subjectLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:inputType="text"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/facultyLayout">
-
-        <android.support.design.widget.TextInputEditText
-            android:id="@+id/subjectInputView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/faculty" />
-    </android.support.design.widget.TextInputLayout>
-
-    <Button
-        android:id="@+id/searchButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@android:string/search_go"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/subjectLayout" />
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/lessonsListView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layoutManager="LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/searchButton" />
-
-</android.support.constraint.ConstraintLayout>
+    android:orientation="vertical"
+    tools:context="pl.edu.zut.mad.schedule.search.SearchActivity" />
