+++ src/main/res/layout/fragment_search.xml
@@ -1,19 +1,16 @@
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
     android:padding="10dp"
-    tools:context="pl.edu.zut.mad.schedule.search.SearchByTeacherFragment">
+    tools:context="pl.edu.zut.mad.schedule.search.SearchActivity">
 
     <android.support.design.widget.TextInputLayout
         android:id="@+id/teacherNameLayout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toTopOf="@id/teacherSurnameLayout"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        android:layout_height="wrap_content">
 
         <android.support.design.widget.TextInputEditText
             android:id="@+id/teacherNameInputView"
@@ -26,10 +23,7 @@
     <android.support.design.widget.TextInputLayout
         android:id="@+id/teacherSurnameLayout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/teacherNameLayout">
+        android:layout_height="wrap_content">
 
         <android.support.design.widget.TextInputEditText
             android:id="@+id/teacherSurnameInputView"
@@ -43,25 +37,45 @@
         android:id="@+id/facultyLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:inputType="text"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/teacherSurnameLayout">
+        android:inputType="text">
 
         <android.support.design.widget.TextInputEditText
-            android:id="@+id/facultyInputView"
+            android:id="@+id/facultyAbbreviationInputView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="@string/faculty" />
     </android.support.design.widget.TextInputLayout>
 
+    <android.support.design.widget.TextInputLayout
+        android:id="@+id/roomLayout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:inputType="text">
+
+        <android.support.design.widget.TextInputEditText
+            android:id="@+id/roomInputView"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="@string/room" />
+    </android.support.design.widget.TextInputLayout>
+
+    <android.support.design.widget.TextInputLayout
+        android:id="@+id/subjectLayout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:inputType="text">
+
+        <android.support.design.widget.TextInputEditText
+            android:id="@+id/subjectInputView"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="@string/subject" />
+    </android.support.design.widget.TextInputLayout>
+
     <Button
         android:id="@+id/searchButton"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@android:string/search_go"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/facultyLayout" />
+        android:text="@android:string/search_go" />
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
